backend/requirements.txt
flask
flask_cors
openai-python
face_recognition
python-telegram-bot
sqlalchemy
